TrueSight/ai-model/requirements.txt
flask
tensorflow
numpy
pillow
